Fixes and Improvements

+++ tests/test_gini.py
@@ -11,7 +11,7 @@ def test_gini_empty():
     '''test calculating a gini coefficient with an empty list
     This will cause some warnings from python
     '''
-    gini = calculate_gini(pd.Series([]))
+    gini = calculate_gini(pd.Series([], dtype=float))
     assert math.isnan(gini) is True
 
 
@@ -21,7 +21,6 @@ def test_gini_single():
     '''
     gini = calculate_gini(pd.Series([1.0]))
     assert gini == 0.0
-
 
 def test_gini_four():
     '''test calculating a gini coefficient with four different items'''
